+++ CHANGELOG.md
@@ -1,11 +1,13 @@
 # CHANGELOG
 
-## [MARIO/RED-10-AUTH-2] - 2023-04-21
+## [MARIO/RED-10-AUTH] - 2023-04-22
 
 - Crear clase BaseStore.
 - Crear State/Store para Report (incluye Wood y "Reports") y User (incluye License y Company).
 - Ajustar los servicios para que devuelvan y escriban el state apropiadamente.
 - Agregar modelos para WtCompany y WtUser.
+- Make UserService reactive to authentication state change.
+- Make Analysis page reactive.
 
 ## [MARIO/RED-10-AUTH] - 2023-04-21